requirements.txt
# Core LLM library - build from source on Raspberry Pi 5
llama-cpp-python>=0.2.0

# Terminal UI for rich display and cyberpunk aesthetics
rich>=13.0.0

# System monitoring for resource utilization
psutil>=5.9.0

# Networking and JSON handling
requests>=2.28.0

# Build dependencies for compiling llama-cpp-python from source
cmake>=3.20.0
pybind11>=2.10.0

# Optional: Better performance profiling
py-cpuinfo>=9.0.0

# Development and testing
pytest>=7.0.0

# Visual capabilities
opencv-python>=4.8.0
picamera2>=0.3.12
pillow>=10.0.0

# Conversation logging and replay
# sqlite3 is built into Python standard library

# Web server and monitoring interface
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
pyjwt>=2.8.0
python-socketio>=5.10.0
eventlet>=0.33.0

# Statistical analysis and data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
kaleido>=0.2.1  # For plotly static image export
networkx>=3.0  # Network graph analysis and visualization
jupyter>=1.0.0  # Jupyter notebook support
ipywidgets>=8.0.0  # Interactive widgets for Jupyter

# Interactive dashboard
streamlit>=1.28.0  # Web-based dashboard framework
tabulate>=0.9.0  # Table formatting for markdown export